api/requirements.txt
annotated-types==0.7.0
anthropic==0.74.1
anyio==4.11.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
azure-functions==1.24.0
bcrypt==4.2.1
blinker==1.9.0
brotli==1.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
distro==1.9.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
freetype-py==2.5.1
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
lxml==6.0.2
MarkupSafe==3.0.3
msal==1.34.0
numpy==2.3.4
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydyf==0.11.0
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pypdf==6.3.0
pyphen==0.17.2
pyspnego==0.12.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.3
reportlab==4.4.5
requests==2.32.5
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
rlPyCairo==0.4.0
sendgrid==6.12.5
SharePlum==0.5.1
six==1.17.0
sniffio==1.3.1
sspilib==0.4.0
svglib==1.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
zopfli==0.4.0
